Buff controller, DamageOverTime, and Flay skill

+++ Hellforge.Game/Assets/Scripts/Hellforge.Game/Entities/IDamageable.cs
@@ -4,7 +4,7 @@ namespace Hellforge.Game.Entities
     public interface IDamageable
     {
         void Damage(DamageInfo dmgInfo);
-        int Health { get; }
-        int MaxHealth { get; }
+        float Health { get; }
+        float MaxHealth { get; }
     }
 }
